reorganize pallet_tools for better pallet separation
