+++ pallets/qa-tools/src/pallet_tools/mod.rs
@@ -33,4 +33,6 @@
 
 pub mod liquidity_proxy;
 pub mod order_book;
+pub mod pool_xyk;
 pub mod price_tools;
+pub mod xst;
